+++ c-ares/CHANGELOG.md
@@ -12,6 +12,8 @@
 - Result iterators now implement `FusedIterator`; `AResultsIter` and
   `AAAAResultsIter` also implement `ExactSizeIterator`
 - Rename `NAPTRResult::reg_exp()` to `regexp()`
+- Rename `get_servers()` to `servers()` and return `Vec<String>` instead of
+  `AresString`
 - Derive Debug where it was previously missing
 - Return errors instead of panicking when string arguments contain null bytes.
   The following methods now return `Result`s: